+++ src/matcher_combinators/printer.clj
@@ -3,14 +3,16 @@
   (:require [clojure.pprint :as pprint]
             [matcher-combinators.model :as model]
             [colorize.core :as colorize])
-  (:import [matcher_combinators.model Mismatch Missing Unexpected FailedPredicate]))
+  (:import [matcher_combinators.model Mismatch Missing Unexpected FailedPredicate InvalidMatcherType]))
 
 (defrecord ColorTag [color expression])
 
 (defmulti markup-expression class)
 
 (defmethod markup-expression Mismatch [mismatch]
-  (list 'mismatch (->ColorTag :yellow (:expected mismatch)) (->ColorTag :red (:actual mismatch))))
+  (list 'mismatch
+        (->ColorTag :yellow (:expected mismatch))
+        (->ColorTag :red (:actual mismatch))))
 
 (defmethod markup-expression Missing [missing]
   (list 'missing (->ColorTag :red (:expected missing))))
@@ -19,7 +21,14 @@
   (list 'unexpected (->ColorTag :red (:actual unexpected))))
 
 (defmethod markup-expression FailedPredicate [failed-predicate]
-  (list 'predicate (->ColorTag :yellow (:form failed-predicate)) (->ColorTag :red (:actual failed-predicate))))
+  (list 'predicate
+        (->ColorTag :yellow (:form failed-predicate))
+        (->ColorTag :red (:actual failed-predicate))))
+
+(defmethod markup-expression InvalidMatcherType [invalid-type]
+  (list 'invalid-matcher-input
+        (->ColorTag :yellow (:expected-type-msg invalid-type))
+        (->ColorTag :red (:provided invalid-type))))
 
 (defmethod markup-expression :default [expression]
   expression)