+++ test/matcher_combinators/core_test.clj
@@ -239,13 +239,37 @@
 
 (let [matchers [(pred-matcher odd?) (pred-matcher even?)]]
   (fact "subset will recur on matchers"
-    (#'core/matches-in-any-order? matchers [5 4 1 2] true) => truthy
-    (#'core/matches-in-any-order? matchers [5 1 3 2] true) => falsey)
+    (#'core/matches-in-any-order? matchers [5 4 1 2] true [])
+    => (contains {:matched?  true
+                  :unmatched empty?
+                  :matched   vector?})
+    (#'core/matches-in-any-order? matchers [5 1 3 2] true [])
+    => (contains {:matched?  false
+                  :unmatched (just [anything])
+                  :matched   (just [anything])}))
   (fact "works well with identical matchers"
-    (#'core/matches-in-any-order? [(equals-value 2) (equals-value 2)] [2 2] false)
-    => truthy)
+    (#'core/matches-in-any-order? [(equals-value 2) (equals-value 2)] [2 2] false [])
+    => (contains {:matched?  true
+                  :unmatched empty?
+                  :matched   (just [anything anything])}))
   (fact "mismatch if there are more matchers than actual elements"
     (#'core/match-any-order matchers [5] false)
     => [:mismatch (model/->Mismatch matchers [5])]
     (#'core/match-any-order matchers [5] true)
     => [:mismatch (model/->Mismatch matchers [5])]))
+
+(tabular
+  (fact "Providing seq/map matcher with incorrect input leads to automatic mismatch"
+    (core/match (?matcher 1) 1)
+    => (just [:mismatch
+              (contains {:expected-type-msg
+                         #(clojure.string/starts-with? % (-> ?matcher var meta :name str))
+
+                         :provided
+                         "provided: 1"})]))
+  ?matcher
+  equals-seq
+  sublist
+  subset
+  equals-map
+  contains-map)